[WIP] composer "extra config package" based bundle registration
Every library package that has configured the sepcified key in its
`extra` config will be registered automatically:

``` json

{
    "name": "vendor/foo-bundle"
    "extra": {
        "register-bundles": ["Vendor\Bundle\VendorFooBundle"]
    }
}

```

The effect is to create a file in `app/bundle/vendor/foo-bundle` that
contains php code to instanciate the bundle:

``` php

<?php return new \Vendor\Bundle\VendorFooBundle($this);

```

The bundle awlays receive the kernel as first and only constructor argument.

A Kernel impl. may then be able to auto-discover and `require` these
files to append them to its list of registered bundles.